Create orderGene.py
To make sure the ligand~receptor gene pair have the same order (e.g. avoid receptor~ligand pairs) as mentioned in https://github.com/Teichlab/cellphonedb/issues/214, this script could be used to exchange the gene pairs where A is a receptor and B is not a receptor in interaction_input.csv.
After that, the receptor~ligand pairs will be changed into ligand~receptor pairs.